requirements.txt
Django>=4.2,<5.0
django-cors-headers>=4.3.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-filter>=21.0,<25.0
psycopg2-binary>=2.9
python-decouple>=3.8
requests>=2.28
razorpay>=1.4
gunicorn>=21.2
whitenoise>=6.5
dj-database-url>=2.1

# ── Community Intelligence Engine ────────────────────────────────────────
anthropic>=0.40.0
# Optional — enables better topic clustering. Pipeline falls back to a
# hashed bag-of-words embedding when missing.
# sentence-transformers>=2.7.0
# Optional — enables yt-dlp-based YouTube trending fallback.
# yt-dlp>=2024.4.9
